requirements.txt
celery
cryptography
Django>=1.11,<2.0
django-allauth
django-bootstrap-form
django-celery-beat
django-celery-results
django-debug-toolbar
djangorestframework
drf-nested-routers
petlib
Pillow
psycopg2-binary
pytz
redis
reportlab
requests
rules
six
